pyproject.toml
[project]
name = "sidraupdates"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "python-dotenv>=1.2.2",
    "python-telegram-bot>=22.6",
    "requests>=2.32.5",
]
